hems_dental/requirements.txt
Django==4.2.6
django_phonenumber_field==7.2.0
